Refactor lineage graph handling and add utility functions for intermediate nodes management
- Update project version to 0.3.1
- Change development status to Beta in pyproject.toml
- Introduce IntermediateNodeStore class in utils.py for managing intermediate nodes
- Implement filter_intermediate_nodes function to streamline node processing
- Refactor LineageGraph to utilize new utility functions
- Enhance test coverage for IntermediateNodeStore and LineageGraph
