+++ pyproject.toml
@@ -11,7 +11,7 @@ dependencies = [
     "sqlglot>=26.13.0,<27.0.0",
 ]
 classifiers = [
-    "Development Status :: 3 - Alpha",
+    "Development Status :: 4 - Beta",
     "Programming Language :: Python :: 3 :: Only",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
