+++ src/sql2lineage/__init__.py
@@ -1,4 +1,4 @@
 """sql2lineage package."""
 
 __app_name__ = "sql2lineage"
-__version__ = "0.3.0"
+__version__ = "0.3.1"
