+++ tests/test_graph.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import pytest
 
-from sql2lineage.graph import IntermediateNodeStore, LineageGraph
+from sql2lineage.graph import LineageGraph
 from sql2lineage.model import (
     ColumnLineage,
     LineageResult,
@@ -36,10 +36,16 @@ class TestLinageGraph:
         graph.add_table_edges(
             {
                 SourceTable(
-                    source="source_table", target="target_table", alias="alias_name"
+                    source="source_table",
+                    target="target_table",
+                    alias="alias_name",
+                    type="TABLE",
                 ),
                 SourceTable(
-                    source="source_table2", target="target_table2", alias="alias_name2"
+                    source="source_table2",
+                    target="target_table2",
+                    alias="alias_name2",
+                    type="TABLE",
                 ),
             }
         )
@@ -78,10 +84,16 @@ class TestLinageGraph:
         graph.add_table_edges(
             {
                 SourceTable(
-                    source="source_table", target="target_table", alias="alias_name"
+                    source="source_table",
+                    target="target_table",
+                    alias="alias_name",
+                    type="TABLE",
                 ),
                 SourceTable(
-                    source="source_table2", target="target_table2", alias="alias_name2"
+                    source="source_table2",
+                    target="target_table2",
+                    alias="alias_name2",
+                    type="TABLE",
                 ),
             }
         )
@@ -400,44 +412,23 @@ class TestLinageGraph:
             ]
         ]
 
+    def test_print_neighbourhood(self, graph, capsys):
+        """Test print neighbourhood."""
+        parser = SQLLineageParser(dialect="bigquery")
 
-class TestIntermediateNodeStore:
-    """Test IntermediateNodeStore."""
+        with Path("tests/sql/example.sql").open("r", encoding="utf-8") as src:
+            parsed_result = parser.extract_lineage(src.read())
 
-    def test_setitem_and_getitem(self):
-        """Test that using __setitem__ and __getitem__ works correctly."""
-        store = IntermediateNodeStore()
-        store["node1"] = "value1"
-        store["node2"] = "value2"
-        assert store["node1"] == "value1"
-        assert store["node2"] == "value2"
+        graph = LineageGraph()
+        graph.from_parsed(parsed_result.expressions)
 
-    def test_getitem_keyerror(self):
-        """Test that __getitem__ raises a KeyError for missing keys."""
-        store = IntermediateNodeStore()
-        store["node1"] = "value1"
-        with pytest.raises(KeyError):
-            _ = store["non_existing_node"]
+        graph.print_neighbourhood(
+            graph.get_node_neighbours(node="big_orders", node_type="TABLE")
+        )
+        captured = capsys.readouterr()
+        expected = """Neighbourhood:
+  ↳ {'source': 'raw.orders', 'target': 'orders_with_tax', 'type': 'TABLE', 'table_type': 'CTE'}
+  ↳ {'source': 'orders_with_tax', 'target': 'filtered_orders', 'type': 'TABLE', 'table_type': 'CTE'}
+  ↳ {'source': 'filtered_orders', 'target': 'big_orders', 'type': 'TABLE', 'table_type': 'TABLE'}"""
 
-    def test_contains(self):
-        """Test that __contains__ correctly identifies existing and non-existing keys."""
-        store = IntermediateNodeStore()
-        store["node1"] = "value1"
-        store.add(("node2", "value2"))
-        assert "node1" in store
-        assert "node2" in store
-        assert "non_existing" not in store
-
-    def test_add_method(self):
-        """Test that the add() method correctly adds nodes."""
-        store = IntermediateNodeStore()
-        store.add(("node3", "value3"))
-        assert store["node3"] == "value3"
-
-    def test_duplicate_keys(self):
-        """Test that duplicate keys return the first inserted value when using __getitem__."""
-        store = IntermediateNodeStore()
-        store["node_dup"] = "first_value"
-        store.add(("node_dup", "second_value"))
-        # __getitem__ should return the value from the first occurrence.
-        assert store["node_dup"] == "first_value"
+        assert expected in captured.out
